magit-insert-unpushed-to-upstream-or-recent: new function
This makes a lot more sense than
`magit-insert-unpulled-from-upstream-or-recent'.  Show all the
new commits that haven't made it upstream yet, or if there are no
such commits, then show a bunch of recent commits.  When using this
function then there always is a list of recent local commits (which
may or may not have been pushed upstream yet).